pybtex dependency, remove_field command
I added pybtex as a dependency to make it easier to write and parse
bibtex files in a structured way.

I added a remove_field command that uses pybtex to remove a given
field.

This really isn't in keeping with the state purpose of bibby and maybe
it should just be a script but it is in here anyhow for the time being